Split chunked wfifo and rfifo macroses into separate files
