+++ src/api/HPMapi.c
@@ -25,7 +25,7 @@
 #include "common/cbasetypes.h"
 
 #include "common/HPMi.h"
-#include "common/chunked.h"
+#include "common/chunked/rfifo.h"
 #include "common/conf.h"
 #include "common/console.h"
 #include "common/core.h"
