+++ src/api/handlers.c
@@ -25,7 +25,7 @@
 
 #include "common/cbasetypes.h"
 #include "common/api.h"
-#include "common/chunked.h"
+//#include "common/chunked.h"
 #include "common/memmgr.h"
 #include "common/nullpo.h"
 #include "common/showmsg.h"
